Bot PoC works, updated .gitignore
Added local/ scratch directory for private data such as auth keys.

Found working examples on the 'rework' branch of the Discord.hs repo
<https://github.com/jano017/Discord.hs/tree/rework>
Followed the ping-pong example to understand how the discord-hs API
works. Personal opinion: lots of cool type-level stuff.

Not in this commit: also created the app and an associated bot
account on Discord's side. Keys are in the `local/` directory
and thus aren't committed. The current code expects an auth token
in `local/token.txt`.